release-wes001 Corrigido models, reposity, services e resources e integração com bd

+++ src/main/java/com/duodev/duodevbackend/repository/MentoriaRepository.java
@@ -3,5 +3,5 @@ package com.duodev.duodevbackend.repository;
 import com.duodev.duodevbackend.model.Mentoria;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MentoriaRepository extends JpaRepository<Mentoria, Long> {
+public interface MentoriaRepository extends JpaRepository<Mentoria, Integer> {
 }
