+++ src/main/java/com/duodev/duodevbackend/repository/UsuarioAreaInteresseRepository.java
@@ -3,5 +3,5 @@ package com.duodev.duodevbackend.repository;
 import com.duodev.duodevbackend.model.UsuarioAreaInteresse;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UsuarioAreaInteresseRepository extends JpaRepository<UsuarioAreaInteresse, Long> {
+public interface UsuarioAreaInteresseRepository extends JpaRepository<UsuarioAreaInteresse, Integer> {
 }
